record gpu property with True or False when it's deterministic
