+++ mars/core.py
@@ -667,6 +667,16 @@ class TileableOperandMixin(object):
     def check_inputs(self, inputs):
         pass
 
+    def _check_if_gpu(self, inputs):
+        if getattr(self, '_gpu', None) is None and \
+                inputs is not None and len(inputs) > 0:
+            if all(inp.op.gpu is True for inp in inputs):
+                self._gpu = True
+            elif all(inp.op.gpu is False for inp in inputs):
+                self._gpu = False
+            else:
+                self._gpu = None  # not sure if put on GPU
+
     def _create_chunk(self, output_idx, index, **kw):
         raise NotImplementedError
 
@@ -677,6 +687,7 @@ class TileableOperandMixin(object):
 
         self.check_inputs(inputs)
         getattr(self, '_set_inputs')(inputs)
+        self._check_if_gpu(self._inputs)
         if getattr(self, '_key', None) is None:
             getattr(self, '_update_key')()
 
@@ -730,6 +741,7 @@ class TileableOperandMixin(object):
 
         self.check_inputs(inputs)
         getattr(self, '_set_inputs')(inputs)
+        self._check_if_gpu(self._inputs)
         if getattr(self, '_key', None) is None:
             getattr(self, '_update_key')()  # update key when inputs are set
