+++ mars/dataframe/datasource/from_tensor.py
@@ -132,5 +132,6 @@ def from_tensor(tensor, index=None, columns=None, gpu=None, sparse=False):
         col_num = tensor.shape[1]
     except IndexError:
         col_num = 1
+    gpu = tensor.op.gpu if gpu is None else gpu
     op = DataFrameFromTensor(dtypes=pd.Series([tensor.dtype] * col_num), gpu=gpu, sparse=sparse)
     return op(tensor, index, columns)
